+++ Examples.Common/_Imports.razor
@@ -6,4 +6,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
 @using CodeMirror6
+@using CodeMirror6.Commands
 @using CodeMirror6.Models
